more overflow possibilities

+++ server/src/main/java/org/elasticsearch/index/fielddata/SortingNumericDocValues.java
@@ -68,7 +68,7 @@ public abstract class SortingNumericDocValues extends SortedNumericDocValues {
         // to include both the additional bytes used by the grown array
         // as well as the overhead of keeping both arrays in memory while
         // copying.
-        long oldValuesSizeInBytes = values.length * Long.BYTES;
+        long oldValuesSizeInBytes = (long) values.length * Long.BYTES;
         int newValuesLength = ArrayUtil.oversize(newSize, Long.BYTES);
         circuitBreakerConsumer.accept((long) newValuesLength * Long.BYTES);
